+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - Raised minimum PHP version required to 5.6.0 or 7.2.0.
+- Updated minimum PHPUnit version to 5.2.0 for unit tests.
+- PHPUnit 8.5 and 9.0 can now be used to run the unit tests.
+- PHPUnit 6 and 7 are no longer supported.
 
 ## [v1.11.0] - 2020-02-14